pyproject.toml
[project]
name = "service-maker"
version = "0.1.0"
description = ""
authors = [
    {name = "Leo Hehlen",email = "leo.hehlen@pc-rtx.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs (>=25.3.0,<26.0.0)"
]

[tool.poetry]
packages = [{include = "service_maker", from = "src"}]

[tool.poetry.scripts]
service-maker = "service_maker.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
